requirements.txt
python-dotenv
pysnmp
requests
python-telegram-bot
